Move MetricController to its own file
